+++ docs/src/changelog.md
@@ -46,6 +46,7 @@ project adheres to [Semantic Versioning](https://semver.org/).
     - Searching entries
 - Rework edition forms as modals:
     - Bookmark tag edition form
+    - Feed category edition form
 
 ## [v0.6.1](https://github.com/virtualtam/sparklemuffin/releases/tag/v0.6.1) - 2026-04-23
 ### Security